create skeleton user report view
this view is just a box that appears over GoogleMap. It has a placeholder for an image the user will eventually be able to take, some buttons for all recordable weather conditions, and a submit button at the bottom which, for now, just lowers the user report screen

moved MapScreen into a sub package to group it with a data only relevent to MapScreen, that's why other files changed.

slight problem with this commit, the interactivity with the weather type buttons is there, but doesn't seem to show when demoing, only when previewing. will figure out why later.